feat(utxo-staking): move assertion utils to fixtures.utils.ts
Move assertion utility functions from transactions.ts to fixtures.utils.ts
for better code organization and reusability.

Issue: BTC-2143
