+++ modules/utxo-staking/test/unit/fixtures.utils.ts
@@ -1,6 +1,9 @@
-import { Descriptor } from '@bitgo/wasm-miniscript';
+import assert from 'assert';
+
+import { ast, Descriptor, Miniscript } from '@bitgo/wasm-miniscript';
 import * as utxolib from '@bitgo/utxo-lib';
 import * as bitcoinjslib from 'bitcoinjs-lib';
+import { getFixture } from '@bitgo/utxo-core/testutil';
 import { toPlainObjectFromTx, toPlainObjectFromPsbt } from '@bitgo/utxo-core/testutil/descriptor';
 
 export function normalize(v: unknown): unknown {
@@ -32,4 +35,29 @@ export function normalize(v: unknown): unknown {
     return Object.fromEntries(Object.entries(v).flatMap(([k, v]) => (v === undefined ? [] : [[k, normalize(v)]])));
   }
   return v;
+}
+
+type EqualsAssertion = typeof assert.deepStrictEqual;
+
+export async function assertEqualsFixture(
+  fixtureName: string,
+  value: unknown,
+  n = normalize,
+  eq: EqualsAssertion = assert.deepStrictEqual
+): Promise<void> {
+  value = n(value);
+  eq(await getFixture(fixtureName, value), value);
+}
+
+export async function assertTransactionEqualsFixture(fixtureName: string, tx: unknown): Promise<void> {
+  await assertEqualsFixture(fixtureName, normalize(tx));
+}
+
+export function assertEqualsMiniscript(script: Buffer, miniscript: ast.MiniscriptNode): void {
+  const ms = Miniscript.fromBitcoinScript(script, 'tap');
+  assert.deepStrictEqual(ast.fromMiniscript(ms), miniscript);
+  assert.deepStrictEqual(
+    script.toString('hex'),
+    Buffer.from(Miniscript.fromString(ast.formatNode(miniscript), 'tap').encode()).toString('hex')
+  );
 }
